mm-video-v4l2: vdec: Fix unsigned underflow
- Expand unsigned 32 bit variables to signed 64 bit to
avoid unsigned underflow

+++ mm-video-v4l2/vidc/vdec/inc/omx_vdec.h
@@ -261,7 +261,7 @@ public:
     }
 
     OMX_U32 get_average(OMX_U32 sample) {
-        if (avg && abs(avg - sample) > FPS_RESET_THRESHOLD) {
+        if (avg && abs((long long)avg - sample) > FPS_RESET_THRESHOLD) {
             reset();
         }
